Update TodoList.html
This is a simple To-Do List web application built using HTML, CSS, and JavaScript. The app allows users to add tasks, display them in a list, and remove tasks by clicking on the list items. It is designed to be a basic and easy-to-understand application to showcase how a to-do list can be implemented using plain web technologies.